add components mapp and fix some html and css and little of java
