+++ our-menu.html
@@ -22,10 +22,10 @@
                 <article>
                     <h1 class="menu-text">MENY</h1>
                     <nav class="menu-nav">
-                        <ol class="menu-nav_items" id="all">Visa allt</ol>
-                        <ol class="menu-nav_items" id="dishes">Rätter</ol>
-                        <ol class="menu-nav_items" id="drinks">Drycka</ol>
-                        <ol class="menu-nav_items" id="sauces">Dipsås</ol>
+                        <ol class="menu-nav_items all-btn">Visa allt</ol>
+                        <ol class="menu-nav_items wonton-btn">Rätter</ol>
+                        <ol class="menu-nav_items drink-btn">Drycka</ol>
+                        <ol class="menu-nav_items dip-btn">Dipsås</ol>
                     </nav>
                 </article>
                 <article class="cart">
